+++ monocle/dummy/change.py
@@ -27,7 +27,8 @@ from monocle.db.db import (
     SimpleFile,
     Commit,
 )
-from monocle.basecrawler import BaseCrawler, RawChange, create_ident
+from monocle.basecrawler import BaseCrawler, RawChange
+from monocle.ident import create_ident
 
 name = "dummy_crawler"
 help = "Dummy crawler to demo a crawler for a dummy code review service"
@@ -111,8 +112,8 @@ class DummyChangeFetcher(BaseCrawler):
             commits=[
                 Commit(
                     sha="12345",
-                    author=create_ident(url, "John"),
-                    committer=create_ident(url, "John"),
+                    author=create_ident(url, "John", []),
+                    committer=create_ident(url, "John", []),
                     authored_at="2020-04-11T07:01:15Z",
                     committed_at="2020-04-11T07:00:15Z",
                     additions=10,
@@ -123,9 +124,9 @@ class DummyChangeFetcher(BaseCrawler):
             repository_prefix="org",
             repository_fullname="org/dummyrepo",
             repository_shortname="dummyrepo",
-            author=create_ident(url, "John"),
-            committer=create_ident(url, "John"),
-            merged_by=create_ident(url, "Zuul"),
+            author=create_ident(url, "John", []),
+            committer=create_ident(url, "John", []),
+            merged_by=create_ident(url, "Zuul", []),
             branch="dummy-feature",
             target_branch="main",
             created_at="2020-04-11T07:00:15Z",
@@ -145,7 +146,7 @@ class DummyChangeFetcher(BaseCrawler):
             _id="cce_1",
             _type="ChangeCreatedEvent",
             created_at="2020-04-11T07:00:15Z",
-            author=create_ident(url, "John"),
+            author=create_ident(url, "John", []),
             repository_prefix="org",
             repository_fullname="org/dummyrepo",
             repository_shortname="dummyrepo",
@@ -154,7 +155,7 @@ class DummyChangeFetcher(BaseCrawler):
             target_branch="main",
             number=1,
             change_id="org/dummyrepo/1",
-            on_author=create_ident(url, "John"),
+            on_author=create_ident(url, "John", []),
             on_created_at="2020-04-11T07:00:15Z",
             changed_files=[SimpleFile(path="README.md")],
             approval=None,