+++ monocle/gerrit/review.py
@@ -35,7 +35,8 @@ from monocle.db.db import (
     Commit,
     change_or_event_to_dict,
 )
-from monocle.basecrawler import BaseCrawler, RawChange, create_ident
+from monocle.basecrawler import BaseCrawler, RawChange
+from monocle.ident import create_ident as ci, IdentsConfig
 
 
 name = "gerrit_crawler"
@@ -48,12 +49,13 @@ class GerritCrawlerArgs(object):
     loop_delay: int
     command: str
     base_url: str
-    repository: str
     db: object
-    insecure: bool
-    login: str
-    password: str
-    prefix: str
+    repository: str
+    idents_config: IdentsConfig
+    insecure: Optional[bool]
+    login: Optional[str]
+    password: Optional[str]
+    prefix: Optional[str]
 
 
 class ReviewesFetcher(BaseCrawler):
@@ -62,13 +64,14 @@ class ReviewesFetcher(BaseCrawler):
 
     def __init__(
         self,
-        base_url,
-        repository_prefix,
-        insecure=False,
-        login=None,
-        password=None,
-        prefix=None,
-    ):
+        base_url: str,
+        repository_prefix: str,
+        insecure: bool,
+        idents_config: IdentsConfig,
+        login: Optional[str] = None,
+        password: Optional[str] = None,
+        prefix: Optional[str] = None,
+    ) -> None:
         self.base_url = base_url
         self.repository_prefix = repository_prefix
         self.insecure = insecure
@@ -79,6 +82,10 @@ class ReviewesFetcher(BaseCrawler):
         if login:
             self.auth = HTTPBasicAuth(login, password)
         self.prefix = prefix
+        self.idents_config = idents_config
+
+    def create_ident(self, url: str, uid: str):
+        return ci(url, uid, self.idents_config)
 
     def convert_date_for_db(self, str_date):
         cdate = datetime.strptime(str_date[:-10], "%Y-%m-%d %H:%M:%S").strftime(
@@ -198,7 +205,7 @@ class ReviewesFetcher(BaseCrawler):
                 "repository_fullname": repository_fullname,
                 "repository_shortname": repository_shortname,
                 "url": url,
-                "author": create_ident(
+                "author": self.create_ident(
                     url,
                     "%s/%s"
                     % (review["owner"].get("name"), review["owner"]["_account_id"]),
@@ -220,7 +227,7 @@ class ReviewesFetcher(BaseCrawler):
                 # Note(fbo): Only one assignee possible by review on Gerrit
                 "assignees": (
                     [
-                        create_ident(
+                        self.create_ident(
                             url,
                             "%s/%s"
                             % (
@@ -271,7 +278,7 @@ class ReviewesFetcher(BaseCrawler):
                 _commit["commit"]["committer"]["date"]
             )
             obj["author"] = change["author"]
-            obj["committer"] = create_ident(
+            obj["committer"] = self.create_ident(
                 url,
                 "%s/%s"
                 % (
@@ -303,7 +310,7 @@ class ReviewesFetcher(BaseCrawler):
             if change["state"] == "MERGED":
                 if "submitter" in review:
                     # Gerrit 2.x seems to not have that submitter attribute
-                    change["merged_by"] = create_ident(
+                    change["merged_by"] = self.create_ident(
                         url,
                         "%s/%s"
                         % (
@@ -362,7 +369,7 @@ class ReviewesFetcher(BaseCrawler):
                         "_type": "ChangeCommitPushedEvent",
                         "_id": comment["id"],
                         "created_at": self.convert_date_for_db(comment["date"]),
-                        "author": create_ident(
+                        "author": self.create_ident(
                             url,
                             "%s/%s"
                             % (
@@ -385,7 +392,7 @@ class ReviewesFetcher(BaseCrawler):
                         "_type": "ChangeCommentedEvent",
                         "_id": comment["id"],
                         "created_at": self.convert_date_for_db(comment["date"]),
-                        "author": create_ident(
+                        "author": self.create_ident(
                             url,
                             "%s/%s"
                             % (
@@ -404,9 +411,11 @@ class ReviewesFetcher(BaseCrawler):
                         "_id": "approval_%s" % comment["id"],
                         "created_at": self.convert_date_for_db(comment["date"]),
                         "approval": [
-                            approval_match.groupdict().get("approval").strip()
+                            a.strip()
+                            for a in [approval_match.groupdict().get("approval")]
+                            if a
                         ],
-                        "author": create_ident(
+                        "author": self.create_ident(
                             url,
                             "%s/%s"
                             % (
@@ -462,7 +471,8 @@ if __name__ == "__main__":
     rf = ReviewesFetcher(
         args.base_url,
         args.repository,
-        insecure=args.insecure,
+        args.insecure,
+        idents_config=[],
         login=args.login,
         password=args.password,
         prefix=args.prefix,