+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - [web] Support for REMOTE_USER header to enable external authentication system
+- [config] Added configuration section to enable the merge of contributor's metrics
+  across code review systems.
+- [dbmanage] Added update-idents option.
 
 ### Changed
 ### Removed